Clut (#15)
* Clut

* Update impl/src/main/java/wtf/file/api/util/NumberUtil.java

Fix wrong calulcation


---------

+++ impl/src/main/java/wtf/file/api/v1/decoding/clut/ClutInformation.java
@@ -5,8 +5,8 @@ import wtf.file.api.color.channel.ColorChannel;
 import java.util.Map;
 
 public record ClutInformation(
-        byte codeLength,
-        Map<Integer, Map<ColorChannel, Integer>> clut
+        int codeLength,
+        Map<Long, Map<ColorChannel, Short>> clut
 ) {
 
 }
